Move exit command to a separate package
